refactor: move history from sidebar tab to command palette overlay
- Remove History tab from sidebar (back to Sessions / Runners only)
- Create HistoryCommandPalette using cmdk CommandDialog component
  - Searchable by session name, path, runner name
  - Date-grouped results (Today, Yesterday, This Week, etc.)
  - Resume button for sessions with online runners
  - Load more pagination
- Add Cmd/Ctrl+H keyboard shortcut to toggle history palette
- Add clock icon button in desktop header toolbar
- Add 'Session history' item in user dropdown (desktop + mobile)
- Clean up all showHistory state from sidebar and App.tsx
